Refactor: Move UI-related files from infrastructure to foundation layer
This commit restructures the project by moving UI-related files, specifically `RootAppWidget`, `RootBlocsProvider`, and theme-related files, from the `infrastructure` layer to a new `foundation` layer.

The following files were moved:
- `bricks/flutter_starter_brick/__brick__/lib/shared/widgets/src/root_app_widget.dart` to `bricks/flutter_starter_brick/__brick__/lib/app/src/root_app_widget.dart`
- `bricks/flutter_starter_brick/__brick__/lib/shared/widgets/src/root_blocs_provider.dart` to `bricks/flutter_starter_brick/__brick__/lib/app/src/root_blocs_provider.dart`
- `bricks/flutter_starter_brick/__brick__/lib/infrastructure/ui/theme/src/{{proj_name.snakeCase()}}_theme.dart` to `bricks/flutter_starter_brick/__brick__/lib/foundation/ui/theme/src/{{proj_name.snakeCase()}}_theme.dart`

Additionally, template files for new features under `bricks/flutter_starter_brick/__brick__/bricks/new_feature_brick/` were removed as part of this refactoring.
