--- bricks/flutter_starter_brick/__brick__/bricks/new_feature_brick/__brick__/{{=<% %>=}}{{<%={{ }}=%>feature_name.snakeCase(){{=<% %>=}}}}<%={{ }}=%>/{{=<% %>=}}{{<%={{ }}=%>feature_name.snakeCase(){{=<% %>=}}}}<%={{ }}=%>.dart
@@ -1,1 +1,0 @@
-export '{{=<% %>=}}{{<%={{ }}=%>feature_name.snakeCase(){{=<% %>=}}}}<%={{ }}=%>_screen/{{=<% %>=}}{{<%={{ }}=%>feature_name.snakeCase(){{=<% %>=}}}}<%={{ }}=%>_screen.dart';
