--- bricks/flutter_starter_brick/__brick__/bricks/new_feature_brick/__brick__/{{=<% %>=}}{{<%={{ }}=%>feature_name.snakeCase(){{=<% %>=}}}}<%={{ }}=%>/{{=<% %>=}}{{<%={{ }}=%>feature_name.snakeCase(){{=<% %>=}}}}<%={{ }}=%>_screen/src/ui/{{=<% %>=}}{{<%={{ }}=%>feature_name.snakeCase(){{=<% %>=}}}}<%={{ }}=%>_screen.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:{{proj_name}}/infrastructure/ui/widgets/widgets.dart';
-
-class {{=<% %>=}}{{<%={{ }}=%>feature_name.pascalCase(){{=<% %>=}}}}<%={{ }}=%>Screen extends StatelessWidget {
-  const {{=<% %>=}}{{<%={{ }}=%>feature_name.pascalCase(){{=<% %>=}}}}<%={{ }}=%>Screen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const RootScreenWidget(
-      body: Center(
-        child: Text(
-          '{{=<% %>=}}{{<%={{ }}=%>feature_name.pascalCase(){{=<% %>=}}}}<%={{ }}=%> Screen',
-          style: TextStyle(
-            fontSize: 40,
-          ),
-        ),
-      ),
-    );
-  }
-}
